110920-1950 Commit Mobile (IPhone) App
Checked functionality of Mobile (iPhone) Webapp on test results.
RADAR : Main functionality all working, but the "Go' button on the
virtual keyboard appears to be inoperative. This is may be a problem
with the jQTouch and jQuery libraries (JQTouch is no longer maintained)
and a wholesale re-write of the app may be required. Priority: MEDIUM